+++ pierky/arouteserver/peering_db.py
@@ -66,14 +66,28 @@ class PeeringDBInfo(CachedObject):
 
     @staticmethod
     def _read_from_url(url):
+        PEERING_DB_API_KEY_ENV_VAR = "SECRET_PEERINGDB_API_KEY"
+        PEERING_DB_API_KEY_WELL_KNOWN_FILES = (
+            "~/.arouteserver/peeringdb_api.key",
+            "~/.peeringdb_api.key"
+        )
+
         headers = None
 
         peeringdb_api_key = None
 
-        for env_var in ("SECRET_PEERINGDB_API_KEY", "PEERINGDB_API_KEY", "API_KEY"):
+        for env_var in (PEERING_DB_API_KEY_ENV_VAR, "PEERINGDB_API_KEY", "API_KEY"):
             if env_var in os.environ:
                 peeringdb_api_key = os.environ[env_var].strip()
                 break
+
+        if not peeringdb_api_key:
+            for well_known_file in PEERING_DB_API_KEY_WELL_KNOWN_FILES:
+                path = os.path.expanduser(well_known_file)
+                if os.path.exists(path):
+                    with open(path, "r") as f:
+                        peeringdb_api_key = f.read().strip()
+                        break
 
         if peeringdb_api_key:
             headers = {"Authorization": "Api-Key {}".format(peeringdb_api_key)}
@@ -90,12 +104,18 @@ class PeeringDBInfo(CachedObject):
                 additional_info = ""
                 if e.response.status_code == 429:
                     additional_info = (
-                        " - Please consider setting the SECRET_PEERINGDB_API_KEY "
-                        "environment variable to instruct ARouteServer to use "
-                        "a PeeringDB API key to perform authentication. "
+                        " - Please consider using a PeeringDB API key to perform "
+                        "authentication, which could help mitigating the effects "
+                        "of anonymous API query rate-limit. "
+                        "The key can be configured by setting the environment "
+                        "variable {} or can be stored inside one of the following "
+                        "well-known files: {} "
                         "Documentation on how to create an API key can be found "
                         "on the peeringdb.com web site "
-                        "(https://docs.peeringdb.com/howto/api_keys/)"
+                        "(https://docs.peeringdb.com/howto/api_keys/)".format(
+                            PEERING_DB_API_KEY_ENV_VAR,
+                            ", ".join(PEERING_DB_API_KEY_WELL_KNOWN_FILES)
+                        )
                     )
 
                 raise PeeringDBError(